requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pymongo==4.8.0
python-dotenv==1.0.1
requests==2.32.3
gunicorn==22.0.0
Werkzeug==3.0.3
Jinja2==3.1.4
MarkupSafe==2.1.5
itsdangerous==2.2.0

